Added user CRUD and made some changes in books logic

+++ src/main/java/com/example/library_management_system/Services/BookServices.java
@@ -8,7 +8,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.RequestBody;
 
 import java.util.List;
 import java.util.Optional;
@@ -19,7 +18,7 @@ public class BookServices {
 
     private final BooksRepository booksRepository;
 
-    public ResponseEntity<Book> createBook(@RequestBody Book book) {
+    public ResponseEntity<Book> createBook(Book book) {
         //return ResponseEntity.ok().body(booksRepository.save(book));
         return ResponseEntity.status(HttpStatus.CREATED).body(booksRepository.save(book));
     }
@@ -28,16 +27,32 @@ public class BookServices {
         return ResponseEntity.ok(booksRepository.findAll());
     }
 
-    public ResponseEntity<Optional<Book>> getBookById(int bookId) {
-        return ResponseEntity.ok(booksRepository.findById(bookId));
+    public ResponseEntity<Book> getBookById(int bookId) {
+        Optional<Book> book = booksRepository.findById(bookId);
+        if (book.isPresent()) {
+            return ResponseEntity.ok(book.get());
+        }
+        return ResponseEntity.notFound().build();
     }
 
     public ResponseEntity<Book> updateBook(int id , Book book) {
-        return ResponseEntity.ok(booksRepository.save(book));
+        Optional<Book> optionalBook = booksRepository.findById(id);
+        if (optionalBook.isPresent()) {
+            book.setBookId(id);
+            booksRepository.save(book);
+            return ResponseEntity.ok(book);
+        }
+
+        return ResponseEntity.notFound().build();
     }
 
     public ResponseEntity<Book> deleteBook(int id) {
-        booksRepository.deleteById(id);
-        return  ResponseEntity.ok().build();
+        Optional<Book> optionalBook = booksRepository.findById(id);
+        if (optionalBook.isPresent()) {
+            booksRepository.deleteById(id);
+            return ResponseEntity.ok(optionalBook.get());
+        }
+
+        return ResponseEntity.notFound().build();
     }
 }
